Extract common file system operations
