+++ zork++/src/lib/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod constants;
+pub mod fs;
 pub mod logger;
 pub mod reader;
 pub mod template;
